Allow CSRF configuration, refs #13450
- Add default secret to temporary settings file.
- Render hidden fields and global errors in forms with CSRF protection.
- Add validation to forms without it.
- Use defaults instead of binding GET params before form validation on
  default move action.
- Validate upload limit form sent with Ajax to a different action.
- Fix permissions settings form, use a unique name format on the four
  forms combined on this page to be able to validate each CSRF token.
- Disable CSRF protection in GET forms:
  - Actor's advanced search.
  - IO's advanced search.
  - Description updates.
- Remove not used components:
  - Actor autocomplete search form.
  - QubitWidgetFormSchemaFormatterDetail.

+++ apps/qubit/modules/digitalobject/actions/deleteAction.class.php
@@ -55,29 +55,34 @@ class DigitalObjectDeleteAction extends sfAction
 
     if ($request->isMethod('delete'))
     {
-      // Delete the digital object record from the database
-      $this->resource->delete();
-      QubitSearch::getInstance()->update($this->object);
+      $this->form->bind($request->getPostParameters());
+      
+      if ($this->form->isValid())
+      {
+        // Delete the digital object record from the database
+        $this->resource->delete();
+        QubitSearch::getInstance()->update($this->object);
 
-      if ($this->object instanceOf QubitInformationObject)
-      {
-        $this->object->updateXmlExports();
-      }
-
-      // Redirect to edit page for parent Object
-      if (isset($parent))
-      {
-        $this->redirect(array($parent, 'module' => 'digitalobject', 'action' => 'edit'));
-      }
-      else
-      {
         if ($this->object instanceOf QubitInformationObject)
         {
-          $this->redirect(array($this->object, 'module' => 'informationobject'));
+          $this->object->updateXmlExports();
         }
-        else if ($this->object instanceOf QubitActor)
+
+        // Redirect to edit page for parent Object
+        if (isset($parent))
         {
-          $this->redirect(array($this->object, 'module' => 'actor'));
+          $this->redirect(array($parent, 'module' => 'digitalobject', 'action' => 'edit'));
+        }
+        else
+        {
+          if ($this->object instanceOf QubitInformationObject)
+          {
+            $this->redirect(array($this->object, 'module' => 'informationobject'));
+          }
+          else if ($this->object instanceOf QubitActor)
+          {
+            $this->redirect(array($this->object, 'module' => 'actor'));
+          }
         }
       }
     }
