+++ plugins/qtAccessionPlugin/modules/donor/actions/deleteAction.class.php
@@ -33,14 +33,19 @@ class DonorDeleteAction extends sfAction
 
     if ($request->isMethod('delete'))
     {
-      foreach (QubitRelation::getBySubjectOrObjectId($this->resource->id) as $item)
+      $this->form->bind($request->getPostParameters());
+      
+      if ($this->form->isValid())
       {
-        $item->delete();
+        foreach (QubitRelation::getBySubjectOrObjectId($this->resource->id) as $item)
+        {
+          $item->delete();
+        }
+
+        $this->resource->delete();
+
+        $this->redirect(array('module' => 'donor', 'action' => 'browse'));
       }
-
-      $this->resource->delete();
-
-      $this->redirect(array('module' => 'donor', 'action' => 'browse'));
     }
   }
 }
